Move Stream-K kernel files to a subdirectory
